multi latent phenotypes framework passes some tests
This commit implements the **framework** for multiple
latent phenotypes, but the code still only works for
one latent phenotype as certain parts of the fitting
don't yet handle multiple.

Nonetheless, it has been partially tested to ensure
that:
1. It runs without error.
2. It gives the same results as the previous one-latent
   phenotype framework.

It passes the following tests:
- `tests/monotonicsplineepistasisgaussianlikelihood_model.ipynb`
- `tests/noepistasisgaussianlikelihood_model.ipynb`

The remaining tests still need to be run, and some are expected
to fail simply because of small API changes.