added `MultiLatentSigmoidPhenotypeSimulator`
Allows simulation of libraries with an observed
phenotype that depends on multiple latent phenotypes.

Started a notebook `multi_latent_phenos.ipynb` that
uses this new class to test the fitting of multi-latent
models. But so far the notebook is just implemented
through the testing of the simulation of the data, not
yet the fitting.