refactored `bottleneck_likelihood.ipynb`
The bottleneck likelihood previously had compared
everything to wildtype under the assumption that the
wildtype observed phenotype was zero, but this was nowhere
enforced mathematically. Now enrichments are calculated
related to the mean population fitness, which implicitly
normalizes to whatever is the phenotype of other variants.

This new calculation has now been implemented in
`bottleneck_likelihood.ipynb`, but other parts of
code base relevant (`tests/bottleneck_likelihood_deriv_equations.ipynb`
and the top-line docs in `globalepistasis.py`) still need
to be updated accordingly.